requirements.txt
transformers
datasets
torch
flask
gradio
requests
sentence-transformers
faiss-cpu
streamlit
PyPDF2